modularize sign,ceil,floor,neg
