+++ tfjs-core/src/gradients/Sub_grad.ts
@@ -17,9 +17,9 @@
 import {Sub} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
 import * as broadcast_util from '../ops/broadcast_util';
+import {neg} from '../ops/neg';
 import {reshape} from '../ops/reshape';
 import {sum} from '../ops/sum';
-import {neg} from '../ops/unary_ops';
 import {Tensor} from '../tensor';
 
 export const subGradConfig: GradConfig = {
