+++ tfjs-core/src/kernel_names.ts
@@ -119,6 +119,9 @@ export interface BroadCastToAttrs {
   inputShape: number[];  // for gradient
 }
 
+export const Ceil = 'Ceil';
+export type CeilInputs = UnaryInputs;
+
 export const Complex = 'Complex';
 export type ComplexInputs = Pick<NamedTensorInfoMap, 'real'|'imag'>;
 
@@ -255,6 +258,9 @@ export type EluGradInputs = Pick<NamedTensorInfoMap, 'dy'|'y'>;
 
 export const Equal = 'Equal';
 export type EqualInputs = BinaryInputs;
+
+export const Floor = 'Floor';
+export type FloorInputs = UnaryInputs;
 
 export const FloorDiv = 'FloorDiv';
 export type FloorDivInputs = BinaryInputs;
@@ -404,6 +410,9 @@ export type ModInputs = BinaryInputs;
 export const Multiply = 'Multiply';
 export type MultiplyInputs = BinaryInputs;
 
+export const Negate = 'Negate';
+export type NegateInputs = UnaryInputs;
+
 export const NotEqual = 'NotEqual';
 export type NotEqualInputs = BinaryInputs;
 
@@ -505,6 +514,9 @@ export type SelectV2Inputs = Pick<NamedTensorInfoMap, 'condition'|'t'|'e'>;
 export const Selu = 'Selu';
 export type SeluInputs = Pick<NamedTensorInfoMap, 'x'>;
 
+export const Sign = 'Sign';
+export type SignInputs = UnaryInputs;
+
 export const Sum = 'Sum';
 export type SumInputs = Pick<NamedTensorInfoMap, 'x'>;
 export interface SumAttrs {
@@ -546,6 +558,8 @@ export type TransposeInputs = Pick<NamedTensorInfoMap, 'x'>;
 export interface TransposeAttrs {
   perm: number[];
 }
+
+export type UnaryInputs = Pick<NamedTensorInfoMap, 'x'>;
 
 export const Unpack = 'Unpack';
 export type UnpackInputs = Pick<NamedTensorInfoMap, 'value'>;
