+++ tfjs-core/src/ops/ceil.ts
@@ -1,0 +1,46 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {ENGINE} from '../engine';
+import {Ceil, CeilInputs} from '../kernel_names';
+import {Tensor} from '../tensor';
+import {NamedTensorMap} from '../tensor_types';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+
+import {op} from './operation';
+
+/**
+ * Computes ceiling of input `tf.Tensor` element-wise: `ceil(x)`
+ *
+ * ```js
+ * const x = tf.tensor1d([.6, 1.1, -3.3]);
+ *
+ * x.ceil().print();  // or tf.ceil(x)
+ * ```
+ * @param x The input Tensor.
+ */
+/** @doc {heading: 'Operations', subheading: 'Basic math'} */
+function ceil_<T extends Tensor>(x: T|TensorLike): T {
+  const $x = convertToTensor(x, 'x', 'ceil');
+
+  const inputs: CeilInputs = {x: $x};
+  return ENGINE.runKernelFunc(
+      backend => backend.ceil($x), inputs as {} as NamedTensorMap,
+      null /* grad */, Ceil);
+}
+export const ceil = op({ceil_});
