+++ tfjs-core/src/ops/sigmoid_cross_entropy.ts
@@ -24,11 +24,12 @@ import {add} from './add';
 import {computeWeightedLoss} from './compute_weighted_loss';
 import {Reduction} from './loss_ops_utils';
 import {mul} from './mul';
+import {neg} from './neg';
 import {op} from './operation';
 import {relu} from './relu';
 import {sub} from './sub';
 import {scalar} from './tensor_ops';
-import {abs, exp, log1p, neg} from './unary_ops';
+import {abs, exp, log1p} from './unary_ops';
 
 function sigmoidCrossEntropyWithLogits_<T extends Tensor, O extends Tensor>(
     labels: T|TensorLike, logits: T|TensorLike): O {
