+++ tfjs-core/src/ops/softmax_cross_entropy.ts
@@ -29,12 +29,13 @@ import {div} from './div';
 import {logSumExp} from './log_sum_exp';
 import {Reduction} from './loss_ops_utils';
 import {mul} from './mul';
+import {neg} from './neg';
 import {op} from './operation';
 import {reshape} from './reshape';
 import {sub} from './sub';
 import {sum} from './sum';
 import {scalar} from './tensor_ops';
-import {exp, neg} from './unary_ops';
+import {exp} from './unary_ops';
 
 /**
  * Computes softmax cross entropy between logits and labels.
