+++ tfjs-core/src/ops/unary_ops_test.ts
@@ -253,80 +253,6 @@ describeWithFlags('step', ALL_ENVS, () => {
   });
 });
 
-describeWithFlags('neg', ALL_ENVS, () => {
-  it('basic', async () => {
-    const a = tf.tensor1d([1, -3, 2, 7, -4]);
-    const result = tf.neg(a);
-    expectArraysClose(await result.data(), [-1, 3, -2, -7, 4]);
-  });
-
-  it('propagate NaNs', async () => {
-    const a = tf.tensor1d([1, -3, 2, 7, NaN]);
-    const result = tf.neg(a);
-    const expected = [-1, 3, -2, -7, NaN];
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(4);
-    const dy = tf.scalar(8);
-
-    const da = tf.grad(a => tf.neg(a))(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-    expectArraysClose(await da.data(), [8 * -1]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(4);
-    const dy = tf.scalar(8);
-
-    const da = tf.grad(a => tf.neg(a.clone()).clone())(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-    expectArraysClose(await da.data(), [8 * -1]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([1, 2, -3, 5]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const da = tf.grad(a => tf.neg(a))(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-    expectArraysClose(await da.data(), [1 * -1, 2 * -1, 3 * -1, 4 * -1]);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([3, -1, -2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const da = tf.grad(a => tf.neg(a))(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-    expectArraysClose(await da.data(), [1 * -1, 2 * -1, 3 * -1, 4 * -1]);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.neg({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'neg' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const result = tf.neg([1, -3, 2, 7, -4]);
-    expectArraysClose(await result.data(), [-1, 3, -2, -7, 4]);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.neg('q'))
-        .toThrowError(/Argument 'x' passed to 'neg' must be numeric/);
-  });
-});
-
 describeWithFlags('sigmoid', ALL_ENVS, () => {
   it('basic', async () => {
     const values = [1, -3, 2, 7, -4];
@@ -1227,226 +1153,6 @@ describeWithFlags('log1p', ALL_ENVS, () => {
   it('throws for string tensor', () => {
     expect(() => tf.log1p('q'))
         .toThrowError(/Argument 'x' passed to 'log1p' must be numeric/);
-  });
-});
-
-describeWithFlags('ceil', ALL_ENVS, () => {
-  it('basic', async () => {
-    const a = tf.tensor1d([1.5, 2.1, -1.4]);
-    const r = tf.ceil(a);
-    expectArraysClose(await r.data(), [2, 3, -1]);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([1.5, NaN, -1.4]);
-    const r = tf.ceil(a);
-    expectArraysClose(await r.data(), [2, NaN, -1]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(5.2);
-    const dy = tf.scalar(3);
-
-    const gradients = tf.grad(a => tf.ceil(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(5.2);
-    const dy = tf.scalar(3);
-
-    const gradients = tf.grad(a => tf.ceil(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([-1.1, 2.6, 3, -5.9]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const gradients = tf.grad(a => tf.ceil(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0, 0, 0, 0]);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([-3, 1, 2.2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const gradients = tf.grad(a => tf.ceil(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0, 0, 0, 0]);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.ceil({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'ceil' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const r = tf.ceil([1.5, 2.1, -1.4]);
-    expectArraysClose(await r.data(), [2, 3, -1]);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.ceil('q'))
-        .toThrowError(/Argument 'x' passed to 'ceil' must be numeric/);
-  });
-});
-
-describeWithFlags('floor', ALL_ENVS, () => {
-  it('basic', async () => {
-    const a = tf.tensor1d([1.5, 2.1, -1.4]);
-    const r = tf.floor(a);
-
-    expectArraysClose(await r.data(), [1, 2, -2]);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([1.5, NaN, -1.4]);
-    const r = tf.floor(a);
-    expectArraysClose(await r.data(), [1, NaN, -2]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(5.2);
-    const dy = tf.scalar(3);
-
-    const gradients = tf.grad(a => tf.floor(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(5.2);
-    const dy = tf.scalar(3);
-
-    const gradients = tf.grad(a => tf.floor(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([-1.1, 2.6, 3, -5.9]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const gradients = tf.grad(a => tf.floor(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0, 0, 0, 0]);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([-3, 1, 2.2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const gradients = tf.grad(a => tf.floor(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0, 0, 0, 0]);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.floor({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'floor' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const r = tf.floor([1.5, 2.1, -1.4]);
-    expectArraysClose(await r.data(), [1, 2, -2]);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.floor('q'))
-        .toThrowError(/Argument 'x' passed to 'floor' must be numeric/);
-  });
-});
-
-describeWithFlags('sign', ALL_ENVS, () => {
-  it('basic', async () => {
-    const a = tf.tensor1d([1.5, 0, NaN, -1.4]);
-    const r = tf.sign(a);
-    expectArraysClose(await r.data(), [1, 0, 0, -1]);
-  });
-
-  it('does not propagate NaNs', async () => {
-    const a = tf.tensor1d([1.5, NaN, -1.4]);
-    const r = tf.sign(a);
-    expectArraysClose(await r.data(), [1, 0, -1]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(5.2);
-    const dy = tf.scalar(3);
-
-    const gradients = tf.grad(a => tf.sign(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(5.2);
-    const dy = tf.scalar(3);
-
-    const gradients = tf.grad(a => tf.sign(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([-1.1, 2.6, 3, -5.9]);
-    const dy = tf.tensor1d([-1, 1, 1, -1]);
-
-    const gradients = tf.grad(a => tf.sign(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0, 0, 0, 0]);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([-3, 1, 2.2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const gradients = tf.grad(a => tf.sign(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [0, 0, 0, 0]);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.sign({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'sign' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const r = tf.sign([1.5, 0, NaN, -1.4]);
-    expectArraysClose(await r.data(), [1, 0, 0, -1]);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.sign('q'))
-        .toThrowError(/Argument 'x' passed to 'sign' must be numeric/);
   });
 });
