+++ tfjs-core/src/register_all_gradients.ts
@@ -24,6 +24,7 @@ import {avgPoolGradConfig} from './gradients/AvgPool_grad';
 import {batchMatMulGradConfig} from './gradients/BatchMatMul_grad';
 import {batchToSpaceNDGradConfig} from './gradients/BatchToSpaceND_grad';
 import {broadcastToGradConfig} from './gradients/BroadcastTo_grad';
+import {ceilGradConfig} from './gradients/Ceil_grad';
 import {concatGradConfig} from './gradients/Concat_grad';
 import {conv2DGradConfig} from './gradients/Conv2D_grad';
 import {conv2DBackpropInputGradConfig} from './gradients/Conv2DBackpropInput_grad';
@@ -33,6 +34,7 @@ import {depthwiseConv2dNativeGradConfig} from './gradients/DepthwiseConv2dNative
 import {dilation2dGradConfig} from './gradients/Dilation2D_grad';
 import {divGradConfig} from './gradients/Div_grad';
 import {eluGradConfig} from './gradients/Elu_grad';
+import {floorGradConfig} from './gradients/Floor_grad';
 import {floorDivGradConfig} from './gradients/FloorDiv_grad';
 import {fusedBatchNormGradConfig} from './gradients/FusedBatchNorm_grad';
 import {greaterEqualGradConfig} from './gradients/GreaterEqual_grad';
@@ -46,6 +48,7 @@ import {minGradConfig} from './gradients/Min_grad';
 import {minimumGradConfig} from './gradients/Minimum_grad';
 import {modGradConfig} from './gradients/Mod_grad';
 import {multiplyGradConfig} from './gradients/Multiply_grad';
+import {negateGradConfig} from './gradients/Negate_grad';
 import {oneHotGradConfig} from './gradients/OneHot_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
 import {powGradConfig} from './gradients/Pow_grad';
@@ -58,6 +61,7 @@ import {resizeNearestNeighborGradConfig} from './gradients/ResizeNearestNeighbor
 import {reverseGradConfig} from './gradients/Reverse_grad';
 import {selectV2PoolGradConfig} from './gradients/SelectV2_grad';
 import {seluGradConfig} from './gradients/Selu_grad';
+import {signGradConfig} from './gradients/Sign_grad';
 import {spaceToBatchNDGradConfig} from './gradients/SpaceToBatchND_grad';
 import {splitVGradConfig} from './gradients/SplitV_grad';
 import {squareGradConfig} from './gradients/Square_grad';
@@ -77,14 +81,15 @@ const gradConfigs: GradConfig[] = [
   argMaxGradConfig,
   argMinGradConfig,
   atan2GradConfig,
-  avgPoolGradConfig,
   avgPool3DGradConfig,
+  avgPoolGradConfig,
   batchMatMulGradConfig,
   batchToSpaceNDGradConfig,
   broadcastToGradConfig,
+  ceilGradConfig,
   concatGradConfig,
-  conv2DGradConfig,
   conv2DBackpropInputGradConfig,
+  conv2DGradConfig,
   conv3DGradConfig,
   cumsumGradConfig,
   depthwiseConv2dNativeGradConfig,
@@ -92,39 +97,42 @@ const gradConfigs: GradConfig[] = [
   divGradConfig,
   eluGradConfig,
   floorDivGradConfig,
+  floorGradConfig,
   fusedBatchNormGradConfig,
   greaterEqualGradConfig,
   identityGradConfig,
   lrnGradConfig,
-  oneHotGradConfig,
-  padV2GradConfig,
-  splitVGradConfig,
   maxGradConfig,
-  spaceToBatchNDGradConfig,
   maxGradConfig,
-  minGradConfig,
   maximumGradConfig,
-  maxPoolGradConfig,
   maxPool3DGradConfig,
+  maxPoolGradConfig,
+  minGradConfig,
   minimumGradConfig,
   modGradConfig,
   multiplyGradConfig,
+  negateGradConfig,
   oneHotGradConfig,
+  oneHotGradConfig,
+  padV2GradConfig,
   padV2GradConfig,
   powGradConfig,
   preluGradConfig,
+  relu6GradConfig,
   reluGradConfig,
   reshapeGradConfig,
   resizeBilinearGradConfig,
   resizeNearestNeighborGradConfig,
-  relu6GradConfig,
   reverseGradConfig,
-  seluGradConfig,
   selectV2PoolGradConfig,
+  seluGradConfig,
+  signGradConfig,
+  spaceToBatchNDGradConfig,
   spaceToBatchNDGradConfig,
   splitVGradConfig,
-  squareGradConfig,
+  splitVGradConfig,
   squaredDifferenceGradConfig,
+  squareGradConfig,
   subGradConfig,
   sumGradConfig,
   tileGradConfig,
