+++ tfjs-core/src/tests.ts
@@ -58,6 +58,7 @@ import './ops/binary_ops_test';
 import './ops/boolean_mask_test';
 import './ops/broadcast_to_test';
 import './ops/broadcast_util_test';
+import './ops/ceil_test';
 import './ops/clone_test';
 import './ops/compare_ops_test';
 import './ops/complex_ops_test';
@@ -85,6 +86,7 @@ import './ops/elu_test';
 import './ops/equal_test';
 import './ops/expand_dims_test';
 import './ops/eye_test';
+import './ops/floor_test';
 import './ops/fused_test';
 import './ops/gather_nd_test';
 import './ops/gram_schmidt_test';
@@ -115,6 +117,7 @@ import './ops/moments_test';
 import './ops/moving_average_test';
 import './ops/multi_rnn_cell_test';
 import './ops/multinomial_test';
+import './ops/neg_test';
 import './ops/non_max_suppression_async_test';
 import './ops/non_max_suppression_test';
 import './ops/norm_test';
@@ -142,6 +145,7 @@ import './ops/scatter_nd_test';
 import './ops/segment_ops_test';
 import './ops/selu_test';
 import './ops/sigmoid_cross_entropy_test';
+import './ops/sign_test';
 import './ops/signal_ops_test';
 import './ops/slice_test';
 import './ops/slice_util_test';
